+++ app/src/main/java/eu/kanade/tachiyomi/data/database/models/TrackImpl.kt
@@ -30,6 +30,8 @@ class TrackImpl : Track {
 
     override var private: Boolean = false
 
+    override var custom_lists: String = ""
+
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (javaClass != other?.javaClass) return false