+++ app/src/main/java/eu/kanade/tachiyomi/data/track/anilist/dto/ALUserList.kt
@@ -28,6 +28,7 @@ data class ALUserListItem(
     val startedAt: ALFuzzyDate,
     val completedAt: ALFuzzyDate,
     val private: Boolean = false,
+    val customLists: Map<String, Boolean> = emptyMap(),
     val media: ALSearchItem,
 ) {
     fun toALUserManga(): ALUserManga {
@@ -39,6 +40,7 @@ data class ALUserListItem(
             startDateFuzzy = startedAt.toEpochMilli(),
             completedDateFuzzy = completedAt.toEpochMilli(),
             private = private,
+            customLists = customLists,
             manga = media.toALManga(),
         )
     }